refactor(CA-174): move ProjectSettingsBloc into project_settings feature module
Move the bloc out of libraries and into a proper feature module with
split event/state/bloc files. Remove the dashboard re-export shims and
create ProjectSettingsModule to own the binding.
