+++ lib/features/dashboard/dashboard_module.dart
@@ -2,7 +2,6 @@ import 'package:construculator/app/app_bootstrap.dart';
 import 'package:construculator/app/shell/app_shell_bloc/app_shell_bloc.dart';
 import 'package:construculator/features/dashboard/domain/usecases/watch_recent_estimations_usecase.dart';
 import 'package:construculator/features/dashboard/presentation/bloc/dashboard_bloc/dashboard_bloc.dart';
-import 'package:construculator/features/dashboard/presentation/bloc/project_settings_bloc/project_settings_bloc.dart';
 import 'package:construculator/features/dashboard/presentation/bloc/recent_estimations_bloc/recent_estimations_bloc.dart';
 import 'package:construculator/features/dashboard/presentation/pages/dashboard_page.dart';
 import 'package:construculator/libraries/auth/auth_library_module.dart';
@@ -43,9 +42,6 @@ class DashboardModule extends Module {
         projectRepository: i(),
         currentProjectNotifier: i(),
       ),
-    );
-    i.add<ProjectSettingsBloc>(
-      () => ProjectSettingsBloc(repository: i()),
     );
   }
