--- lib/features/dashboard/presentation/bloc/project_settings_bloc/project_settings_event.dart
@@ -1,1 +1,0 @@
-export 'package:construculator/libraries/project/bloc/project_settings_bloc.dart';
