+++ lib/features/project_settings/presentation/bloc/project_settings_bloc/project_settings_bloc.dart
@@ -5,147 +5,8 @@ import 'package:construculator/libraries/project/domain/repositories/project_set
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-/// Base event for [ProjectSettingsBloc].
-abstract class ProjectSettingsEvent extends Equatable {
-  const ProjectSettingsEvent();
-
-  @override
-  List<Object?> get props => [];
-}
-
-class ProjectSettingsLoadRequested extends ProjectSettingsEvent {
-  final String projectId;
-
-  const ProjectSettingsLoadRequested(this.projectId);
-
-  @override
-  List<Object?> get props => [projectId];
-}
-
-class ProjectSettingsEditingStarted extends ProjectSettingsEvent {
-  final Project project;
-
-  const ProjectSettingsEditingStarted(this.project);
-
-  @override
-  List<Object?> get props => [project];
-}
-
-class ProjectSettingsUpdateSubmitted extends ProjectSettingsEvent {
-  final Project project;
-
-  const ProjectSettingsUpdateSubmitted(this.project);
-
-  @override
-  List<Object?> get props => [project];
-}
-
-class ProjectSettingsDeleteRequested extends ProjectSettingsEvent {
-  final String projectId;
-
-  const ProjectSettingsDeleteRequested(this.projectId);
-
-  @override
-  List<Object?> get props => [projectId];
-}
-
-class ProjectSettingsCreationRequested extends ProjectSettingsEvent {
-  final String name;
-  final String? description;
-  final String? creatorUserId;
-  final StorageProvider? exportStorageProvider;
-
-  const ProjectSettingsCreationRequested({
-    required this.name,
-    required this.creatorUserId,
-    this.description,
-    this.exportStorageProvider,
-  });
-
-  @override
-  List<Object?> get props => [name, description, creatorUserId, exportStorageProvider];
-}
-
-/// Base state for [ProjectSettingsBloc].
-abstract class ProjectSettingsState extends Equatable {
-  const ProjectSettingsState();
-
-  @override
-  List<Object?> get props => [];
-}
-
-class ProjectSettingsInitial extends ProjectSettingsState {
-  const ProjectSettingsInitial();
-}
-
-class ProjectSettingsLoading extends ProjectSettingsState {
-  const ProjectSettingsLoading();
-}
-
-class ProjectSettingsLoaded extends ProjectSettingsState {
-  final Project project;
-
-  const ProjectSettingsLoaded(this.project);
-
-  @override
-  List<Object?> get props => [project];
-}
-
-class ProjectSettingsEditing extends ProjectSettingsState {
-  final Project project;
-  final Project originalProject;
-
-  const ProjectSettingsEditing({
-    required this.project,
-    required this.originalProject,
-  });
-
-  ProjectSettingsEditing copyWith({Project? project, Project? originalProject}) {
-    return ProjectSettingsEditing(
-      project: project ?? this.project,
-      originalProject: originalProject ?? this.originalProject,
-    );
-  }
-
-  @override
-  List<Object?> get props => [project, originalProject];
-}
-
-class ProjectSettingsSaving extends ProjectSettingsState {
-  final Project project;
-
-  const ProjectSettingsSaving(this.project);
-
-  @override
-  List<Object?> get props => [project];
-}
-
-class ProjectSettingsDeleteInProgress extends ProjectSettingsState {
-  const ProjectSettingsDeleteInProgress();
-}
-
-class ProjectCreating extends ProjectSettingsState {
-  const ProjectCreating();
-}
-
-class ProjectCreated extends ProjectSettingsState {
-  final Project project;
-
-  const ProjectCreated(this.project);
-
-  @override
-  List<Object?> get props => [project];
-}
-
-class ProjectSettingsError extends ProjectSettingsState {
-  final Failure failure;
-  final Project? lastProject;
-
-  const ProjectSettingsError({required this.failure, this.lastProject});
-
-  @override
-  List<Object?> get props => [failure, lastProject];
-}
+part 'project_settings_event.dart';
+part 'project_settings_state.dart';
 
 class ProjectSettingsBloc
     extends Bloc<ProjectSettingsEvent, ProjectSettingsState> {
